CameraManagementBackend/requirements.txt
fastapi>=0.109.0,<1
uvicorn>=0.27.0,<1
sqlalchemy>=2.0.36,<3
pg8000>=1.30.4,<2
pydantic>=2.10.0,<3
pydantic-settings>=2.6.0,<3
python-jose>=3.3.0,<4
cryptography>=45.0.7,<47
bcrypt>=4.3.0,<6
python-multipart>=0.0.6,<1
alembic>=1.13.1,<2
